+++ src/packets/ethernetpacket.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-only
 /*
  * TCPPUMP <https://github.com/amartin755/tcppump>
- * Copyright (C) 2012-2021 Andreas Martin (netnag@mailbox.org)
+ * Copyright (C) 2012-2025 Andreas Martin (netnag@mailbox.org)
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -249,7 +249,7 @@ void cEthernetPacket::addSnapHeader (uint32_t oui, uint16_t protocol)
 }
 
 
-void cEthernetPacket::addVlanTag (bool isCTag, unsigned id, unsigned prio, unsigned dei)
+void cEthernetPacket::addVlanTag (bool isCTag, uint16_t id, uint16_t prio, uint16_t dei)
 {
     checkPacketLength (sizeof (vlan_t));
